refactor(theme): expose tokens from theme package
