+++ .pnp.cjs
@@ -215,8 +215,8 @@ const RAW_RUNTIME_STATE =
       "reference": "workspace:ui/design"\
     },\
     {\
-      "name": "@atls-ui/theme-tokens",\
-      "reference": "workspace:ui/theme-tokens"\
+      "name": "@atls-ui/theme",\
+      "reference": "workspace:ui/theme"\
     },\
     {\
       "name": "@atls-utils/use-float",\
@@ -288,7 +288,7 @@ const RAW_RUNTIME_STATE =
     ["@atls-ui-parts/tooltip", ["workspace:ui-parts/tooltip"]],\
     ["@atls-ui-parts/upload", ["virtual:8480592f2ac0c5b00b17bd33d76d0c6d79665a3657c712c5bc5180464a86ac9bcb81d64ef54341b94b31e43db66a2fad906cfc6036c573c170560d657b7f96aa#workspace:ui-parts/upload", "workspace:ui-parts/upload"]],\
     ["@atls-ui/design", ["workspace:ui/design"]],\
-    ["@atls-ui/theme-tokens", ["workspace:ui/theme-tokens"]],\
+    ["@atls-ui/theme", ["workspace:ui/theme"]],\
     ["@atls-utils/use-float", ["virtual:e8b379d238bee0c446fdeaddb9725553fcee91eeb8db9f1f4b47e4e36a5c6c26c07ae1861b7d0a5dd574863bc008f048d25fc793392ceeaade0c9801b921d418#workspace:utils/use-float", "workspace:utils/use-float"]],\
     ["@atls-utils/use-hover", ["virtual:90c3fa37a29dbddd012ae896ef9d136c46ca1d320faf1baf430439ab5aac7e1e7c8ef5281a72424df8d016a8c33c9d5ffe66ee23d2face19f27de3042508900f#workspace:utils/use-hover", "virtual:ab74b0ad21d788b5faaef36e69cfdc113b850f1834ee613c0e84e69c90b4422225458e49681dd5efa10b62df03744b2c5390da900b6092fa9e40fd6605cda940#workspace:utils/use-hover", "workspace:utils/use-hover"]],\
     ["@atls-utils/use-outside-click", ["workspace:utils/use-outside-click"]],\
@@ -639,7 +639,7 @@ const RAW_RUNTIME_STATE =
         "packageLocation": "./.yarn/__virtual__/@atls-ui-admin-theme-virtual-ea79003116/1/ui-admin/theme/",\
         "packageDependencies": [\
           ["@atls-ui-admin/theme", "virtual:90c3fa37a29dbddd012ae896ef9d136c46ca1d320faf1baf430439ab5aac7e1e7c8ef5281a72424df8d016a8c33c9d5ffe66ee23d2face19f27de3042508900f#workspace:ui-admin/theme"],\
-          ["@atls-ui/theme-tokens", "workspace:ui/theme-tokens"],\
+          ["@atls-ui/theme", "workspace:ui/theme"],\
           ["@types/next-themes", null],\
           ["@types/react", "npm:19.1.2"],\
           ["@types/react-dom", null],\
@@ -670,7 +670,7 @@ const RAW_RUNTIME_STATE =
         "packageLocation": "./.yarn/__virtual__/@atls-ui-admin-theme-virtual-844aecbdc2/1/ui-admin/theme/",\
         "packageDependencies": [\
           ["@atls-ui-admin/theme", "virtual:b32a26a2fdb9efe95b0d3c1680d4b91fdda6b5202d080748278d5d418cc8670dced3df8a9ed2862d4501349fbf99cee1cc30a7bdbef4b607caffd1b77c40d6b4#workspace:ui-admin/theme"],\
-          ["@atls-ui/theme-tokens", "workspace:ui/theme-tokens"],\
+          ["@atls-ui/theme", "workspace:ui/theme"],\
           ["@types/next-themes", null],\
           ["@types/react", "npm:19.1.2"],\
           ["@types/react-dom", null],\
@@ -700,7 +700,7 @@ const RAW_RUNTIME_STATE =
         "packageLocation": "./ui-admin/theme/",\
         "packageDependencies": [\
           ["@atls-ui-admin/theme", "workspace:ui-admin/theme"],\
-          ["@atls-ui/theme-tokens", "workspace:ui/theme-tokens"],\
+          ["@atls-ui/theme", "workspace:ui/theme"],\
           ["@types/react", "npm:19.1.2"],\
           ["@vanilla-extract/css", "npm:1.17.1"],\
           ["@vanilla-extract/dynamic", "npm:2.1.2"],\
@@ -1933,7 +1933,7 @@ const RAW_RUNTIME_STATE =
         "packageDependencies": [\
           ["@atls-ui-parts/theme", "virtual:9c83884c7a1bbb8b55eeb55db417f22eaf2ba48e2b1a3840cecc6737a018d63bede31efd8c8fa8a1529cfbb08be4bfc8a1a6ac9b3d61a21277c7e98446e604f8#workspace:ui-parts/theme"],\
           ["@atls-ui-parts/layout", "virtual:dbe1a421eee484ea0bdb9d0589d44d9184c9819d8f5bfcb1a73e3d85271cc45ecf4365a26e00f2a3256c13de48b40272578801688978809175001fb2693bc747#workspace:ui-parts/layout"],\
-          ["@atls-ui/theme-tokens", "workspace:ui/theme-tokens"],\
+          ["@atls-ui/theme", "workspace:ui/theme"],\
           ["@storybook/blocks", "virtual:2ad1ea9ed58ec34034d981b71dc2cc771c4371726bab2fec9d2140e3133cea8ccdc8650d00b83aea848bfc43e82b95d31052474390054d601df6c587082ad33d#npm:8.6.12"],\
           ["@storybook/react", "virtual:90c3fa37a29dbddd012ae896ef9d136c46ca1d320faf1baf430439ab5aac7e1e7c8ef5281a72424df8d016a8c33c9d5ffe66ee23d2face19f27de3042508900f#npm:8.6.12"],\
           ["@types/react", "npm:19.1.2"],\
@@ -1962,7 +1962,7 @@ const RAW_RUNTIME_STATE =
         "packageDependencies": [\
           ["@atls-ui-parts/theme", "virtual:dbe1a421eee484ea0bdb9d0589d44d9184c9819d8f5bfcb1a73e3d85271cc45ecf4365a26e00f2a3256c13de48b40272578801688978809175001fb2693bc747#workspace:ui-parts/theme"],\
           ["@atls-ui-parts/layout", "virtual:dbe1a421eee484ea0bdb9d0589d44d9184c9819d8f5bfcb1a73e3d85271cc45ecf4365a26e00f2a3256c13de48b40272578801688978809175001fb2693bc747#workspace:ui-parts/layout"],\
-          ["@atls-ui/theme-tokens", "workspace:ui/theme-tokens"],\
+          ["@atls-ui/theme", "workspace:ui/theme"],\
           ["@storybook/blocks", "virtual:2ad1ea9ed58ec34034d981b71dc2cc771c4371726bab2fec9d2140e3133cea8ccdc8650d00b83aea848bfc43e82b95d31052474390054d601df6c587082ad33d#npm:8.6.12"],\
           ["@storybook/react", "virtual:90c3fa37a29dbddd012ae896ef9d136c46ca1d320faf1baf430439ab5aac7e1e7c8ef5281a72424df8d016a8c33c9d5ffe66ee23d2face19f27de3042508900f#npm:8.6.12"],\
           ["@types/react", "npm:19.1.2"],\
@@ -1991,7 +1991,7 @@ const RAW_RUNTIME_STATE =
         "packageDependencies": [\
           ["@atls-ui-parts/theme", "workspace:ui-parts/theme"],\
           ["@atls-ui-parts/layout", "virtual:ba0336f2ccf109a54b4b8063d4801935b554fc6f958edd172cb3c73a5a5a49118f16d1cd01b286a3749f851b3f37efaad86c3ceef598318c20b6e7cbecad90d2#workspace:ui-parts/layout"],\
-          ["@atls-ui/theme-tokens", "workspace:ui/theme-tokens"],\
+          ["@atls-ui/theme", "workspace:ui/theme"],\
           ["@storybook/blocks", "virtual:2ad1ea9ed58ec34034d981b71dc2cc771c4371726bab2fec9d2140e3133cea8ccdc8650d00b83aea848bfc43e82b95d31052474390054d601df6c587082ad33d#npm:8.6.12"],\
           ["@storybook/react", "virtual:90c3fa37a29dbddd012ae896ef9d136c46ca1d320faf1baf430439ab5aac7e1e7c8ef5281a72424df8d016a8c33c9d5ffe66ee23d2face19f27de3042508900f#npm:8.6.12"],\
           ["@types/react", "npm:19.1.2"],\
@@ -2104,11 +2104,11 @@ const RAW_RUNTIME_STATE =
         "linkType": "SOFT"\
       }]\
     ]],\
-    ["@atls-ui/theme-tokens", [\
-      ["workspace:ui/theme-tokens", {\
-        "packageLocation": "./ui/theme-tokens/",\
+    ["@atls-ui/theme", [\
+      ["workspace:ui/theme", {\
+        "packageLocation": "./ui/theme/",\
         "packageDependencies": [\
-          ["@atls-ui/theme-tokens", "workspace:ui/theme-tokens"]\
+          ["@atls-ui/theme", "workspace:ui/theme"]\
         ],\
         "linkType": "SOFT"\
       }]\
