+++ ui-parts/theme/package.json
@@ -13,7 +13,7 @@
     "postpack": "rm -rf dist"
   },
   "dependencies": {
-    "@atls-ui/theme-tokens": "workspace:*",
+    "@atls-ui/theme": "workspace:*",
     "clsx": "2.1.1",
     "rainbow-sprinkles": "1.0.0"
   },
