+++ ui/theme/package.json
@@ -1,8 +1,13 @@
 {
-  "name": "@atls-ui/theme-tokens",
+  "name": "@atls-ui/theme",
   "version": "0.0.1",
   "license": "BSD-3-Clause",
   "type": "module",
+  "exports": {
+    "./package.json": "./package.json",
+    ".": "./src/index.ts",
+    "./tokens": "./src/tokens/index.ts"
+  },
   "main": "./src/index.ts",
   "files": [
     "dist"
@@ -20,6 +25,11 @@
         "import": "./dist/index.js",
         "types": "./dist/index.d.ts",
         "default": "./dist/index.js"
+      },
+      "./tokens": {
+        "import": "./dist/tokens/index.js",
+        "types": "./dist/tokens/index.d.ts",
+        "default": "./dist/tokens/index.js"
       }
     },
     "main": "dist/index.js",
